tweak(web): TAM-2394: address AI form builder review feedback
